+++ packages/dreamcord/src/types/message.ts
@@ -1,0 +1,9 @@
+import { APIEmbed } from "@dreamcord/api-types";
+
+export type MessageOptions = {
+  tts?: boolean;
+} & (
+  | { content: string; embeds?: undefined }
+  | { content?: undefined; embeds: APIEmbed[] }
+  | { content: string; embeds: APIEmbed[] }
+);